requirements.txt
python-digitalocean
pytest
pytest-cov
paramiko
